inital reforms : migrate to androidx, button
