+++ app/src/main/java/com/example/toshiba/parkme/activity/AvailableActivity.java
@@ -1,11 +1,12 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
-import android.widget.ListView;
+
+import com.example.toshiba.parkme.R;
 
 public class AvailableActivity extends AppCompatActivity {
