+++ app/src/main/java/com/example/toshiba/parkme/activity/BookActivity.java
@@ -1,21 +1,16 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
-import android.app.ActionBar;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.ActionBarContainer;
-import android.util.Log;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.view.View.OnClickListener;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.AutoCompleteTextView;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import com.example.toshiba.parkme.R;
 
 import java.util.Calendar;
