+++ app/src/main/java/com/example/toshiba/parkme/activity/LoginActivity.java
@@ -1,18 +1,16 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.view.View;
-import android.view.View.OnClickListener;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ProgressBar;
 import android.widget.Toast;
-import android.widget.Toolbar;
 
+import com.example.toshiba.parkme.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
@@ -23,21 +21,9 @@ public class LoginActivity extends AppCompatActivity {
 
 
     private FirebaseAuth mAuth;
-    //private FirebaseAuth.AuthStateListener mAuthListener1;
-
     private EditText mEmailField;
     private EditText mPasswordField;
     private Button submit_button21;
-//
-//    @Override
-//    protected void onStart() {
-//        super.onStart();
-//
-//        if(mAuth.getCurrentUser() !=null){
-//            finish();
-//            startActivity(new Intent(LoginActivity.this,ProfileActivity.class));
-//        }
-//    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -49,19 +35,6 @@ public class LoginActivity extends AppCompatActivity {
         submit_button21 = (Button) findViewById(R.id.submit112);
 
         mAuth = FirebaseAuth.getInstance();
-     /*   mAuthListener = new FirebaseAuth.AuthStateListener() {
-            @Override
-            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
-
-                if (firebaseAuth.getCurrentUser() != null) {
-                    //Intent User Account
-                    startActivity(new Intent(LoginActivity.this,ProfileActivity.class));
-                    finish(); //back button will close the app
-                }
-
-            }
-        };
-*/
         submit_button21.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
