+++ app/src/main/java/com/example/toshiba/parkme/activity/MainActivity.java
@@ -1,11 +1,13 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.Button;
+
+import com.example.toshiba.parkme.R;
 
 public class MainActivity extends AppCompatActivity {
