+++ app/src/main/java/com/example/toshiba/parkme/activity/SAMAvailable.java
@@ -1,9 +1,8 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
-import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View.OnClickListener;
 import android.view.View;
@@ -11,6 +10,7 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.toshiba.parkme.R;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
