+++ app/src/main/java/com/example/toshiba/parkme/activity/SdAvailable.java
@@ -1,8 +1,8 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.view.View.OnClickListener;
@@ -10,6 +10,7 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.example.toshiba.parkme.R;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.database.DataSnapshot;
@@ -18,7 +19,8 @@ import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
-public class RCAvailable extends AppCompatActivity {
+public class SdAvailable extends AppCompatActivity {
+
 
     DatabaseReference databaseReference;
 
@@ -55,8 +57,8 @@ public class RCAvailable extends AppCompatActivity {
         dref.addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
-                av = dataSnapshot.child("Parking Area").child("RC").child("Available").getValue(Integer.class);
-                bk = dataSnapshot.child("Parking Area").child("RC").child("Booked").getValue(Integer.class);
+                av = dataSnapshot.child("Parking Area").child("Sd").child("Available").getValue(Integer.class);
+                bk = dataSnapshot.child("Parking Area").child("Sd").child("Booked").getValue(Integer.class);
 
             }
 
@@ -71,7 +73,7 @@ public class RCAvailable extends AppCompatActivity {
 
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
-                int sama = dataSnapshot.child("Parking Area").child("RC").child("Available").getValue(Integer.class);
+                int sama = dataSnapshot.child("Parking Area").child("Sd").child("Available").getValue(Integer.class);
                 Log.d("TAG", "onvaluereceival: "+ sama);
                 SAM.setText(Integer.toString(sama)+" Slots");
 
@@ -94,7 +96,7 @@ public class RCAvailable extends AppCompatActivity {
         back_button.setOnClickListener((new OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent backIntent = new Intent(RCAvailable.this, ProfileActivity.class);
+                Intent backIntent = new Intent(SdAvailable.this, ProfileActivity.class);
 
                 startActivity(backIntent);
             }
@@ -105,7 +107,7 @@ public class RCAvailable extends AppCompatActivity {
         back1.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent back = new Intent(RCAvailable.this, ListActivity.class);
+                Intent back = new Intent(SdAvailable.this, ListActivity.class);
                 startActivity(back);
             }
         });
@@ -123,12 +125,12 @@ public class RCAvailable extends AppCompatActivity {
                 }else {
                     dref.child("Users").child(current_user.getUid()).child("from").setValue(ftime);
                     dref.child("Users").child(current_user.getUid()).child("to").setValue(ttime);
-                    dref.child("Users").child(current_user.getUid()).child("loc").setValue(3);
-                    dref.child("Parking Area").child("RC").child("Available").setValue(av - 1);
-                    dref.child("Parking Area").child("RC").child("Booked").setValue(bk + 1);
+                    dref.child("Users").child(current_user.getUid()).child("loc").setValue(2);
+                    dref.child("Parking Area").child("Sd").child("Available").setValue(av - 1);
+                    dref.child("Parking Area").child("Sd").child("Booked").setValue(bk + 1);
 
                     Toast.makeText(getApplicationContext(), "Booking Successful!", Toast.LENGTH_LONG).show();
-                    Intent back = new Intent(RCAvailable.this, ProfileActivity.class);
+                    Intent back = new Intent(SdAvailable.this, ProfileActivity.class);
                     startActivity(back);
 
 
@@ -143,9 +145,9 @@ public class RCAvailable extends AppCompatActivity {
 
 //    private FirebaseDatabase firebaseDatabase = FirebaseDatabase.getInstance();
 //    private DatabaseReference rootdatabase = firebaseDatabase.getReference();
-//    private DatabaseReference childreference = rootdatabase.child("Parking Area").child("RC");
+//    private DatabaseReference childreference = rootdatabase.child("Parking Area").child("Sd");
 //
-//    public  TextView RC;
+//    public  TextView Sd;
 //    int av,bk;
 //
 //    FirebaseUser current_user;
@@ -159,17 +161,16 @@ public class RCAvailable extends AppCompatActivity {
 //        Intent intent = getIntent();
 //        final int ftime = intent.getIntExtra("fromtime",0);
 //        final int ttime = intent.getIntExtra("totime",0);
-//        //Toast.makeText(getApplicationContext(),"FROM"+ftime+"TO"+ttime,Toast.LENGTH_LONG).show();
 //
-//        RC = (TextView) findViewById(R.id.availability_status);
+//        Sd = (TextView) findViewById(R.id.availability_status);
 //        current_user = FirebaseAuth.getInstance().getCurrentUser();
 //        dref = FirebaseDatabase.getInstance().getReference();
 //
 //        dref.addValueEventListener(new ValueEventListener() {
 //            @Override
 //            public void onDataChange(DataSnapshot dataSnapshot) {
-//                av = dataSnapshot.child("Parking Area").child("Sd").child("Available").getValue(Integer.class);
-//                bk = dataSnapshot.child("Parking Area").child("Sd").child("Booked").getValue(Integer.class);
+//                av = dataSnapshot.child("Parking Area").child("RC").child("Available").getValue(Integer.class);
+//                bk = dataSnapshot.child("Parking Area").child("RC").child("Booked").getValue(Integer.class);
 //
 //            }
 //
@@ -180,16 +181,22 @@ public class RCAvailable extends AppCompatActivity {
 //        });
 //
 //
+//
 //        Button back_button = (Button) findViewById(R.id.back);
+//
 //
 //        childreference.addValueEventListener(new ValueEventListener() {
 //
 //            @Override
 //            public void onDataChange(DataSnapshot dataSnapshot) {
+//             /*   for(DataSnapshot ds : dataSnapshot.getChildren()){
+//                    int l1slot1= ds.child("lhtc1").child("slot1").getValue(Integer.class);
+//                    int l1slot2= ds.child("lhtc1").child("slot2").getValue(Integer.class);
+//                    int l2slot1= ds.child("lhtc2").child("slot1").getValue(Integer.class);  */
 //
-//                int rca= dataSnapshot.child("Available").getValue(Integer.class);
-//                Log.d("TAG", "onvaluereceival: "+ rca);
-//                RC.setText(Integer.toString(rca)+" Slots");
+//                int sda = dataSnapshot.child("Available").getValue(Integer.class);
+//                Log.d("TAG", "onvaluereceival: "+  sda);
+//                Sd.setText(Integer.toString(sda)+" Slots");
 //
 //            }
 //
@@ -204,41 +211,41 @@ public class RCAvailable extends AppCompatActivity {
 //        });
 //
 //
-//
 //        Button bookandpay= (Button) findViewById(R.id.booknpay);
 //
 //        bookandpay.setOnClickListener(new View.OnClickListener() {
 //            @Override
 //            public void onClick(View v) {
 //
+//
 //                if(av==0) {
 //                    Toast.makeText(getApplicationContext(),"NO SLOTS AVAILABLE FOR BOOKING!", Toast.LENGTH_LONG).show();
 //                }else {
 //                    dref.child("Users").child(current_user.getUid()).child("from").setValue(ftime);
 //                    dref.child("Users").child(current_user.getUid()).child("to").setValue(ttime);
-//                    dref.child("Users").child(current_user.getUid()).child("loc").setValue(3);
-//                    dref.child("Parking Area").child("RC").child("Available").setValue(av - 1);
-//                    dref.child("Parking Area").child("RC").child("Booked").setValue(bk + 1);
+//                    dref.child("Users").child(current_user.getUid()).child("loc").setValue(2);
+//                    dref.child("Parking Area").child("Sd").child("Available").setValue(av - 1);
+//                    dref.child("Parking Area").child("Sd").child("Booked").setValue(bk + 1);
 //
 //                    Toast.makeText(getApplicationContext(), "Booking Successful!", Toast.LENGTH_LONG).show();
-//                    Intent back = new Intent(RCAvailable.this, ProfileActivity.class);
+//                    Intent back = new Intent(SdAvailable.this, ProfileActivity.class);
 //                    startActivity(back);
 //
 //
 //                }
-//
-//
+////                final FirebaseUser current_user;
+////                DatabaseReference dref;
 ////                current_user = FirebaseAuth.getInstance().getCurrentUser();
 ////                dref = FirebaseDatabase.getInstance().getReference();
 ////
-////                dref.child("Users").child(current_user.getUid()).child("from").setValue(ftime);
-////                dref.child("Users").child(current_user.getUid()).child("to").setValue(ttime);
-//
 ////                dref.addValueEventListener(new ValueEventListener() {
 ////                    @Override
 ////                    public void onDataChange(DataSnapshot dataSnapshot) {
-////                     dataSnapshot.child("Users").child(current_user.getUid()).child("from").setValue(ftime);
-////                     //   dataSnapshot.child("Users").child(current_user.getUid()).child("to").getValue(Integer.class);
+////                        int from = dataSnapshot.child("Users").child(current_user.getUid()).child("from").getValue(Integer.class);
+////                        int to = dataSnapshot.child("Users").child(current_user.getUid()).child("to").getValue(Integer.class);
+////
+////                        //receive data from book activity for TO
+////
 ////
 ////                        //from and to data in firebase to be changed to current hour+1 to to_value_from_book_activity
 ////                        //Also decrement the value at parkingarea -> SAM -> Available
@@ -251,8 +258,19 @@ public class RCAvailable extends AppCompatActivity {
 ////
 ////                    }
 ////                });
+////
+////                Toast.makeText(getApplicationContext(),"Your slot is booked! ", Toast.LENGTH_LONG).show();
 //
-//             //   Toast.makeText(getApplicationContext(),"Booking Successful", Toast.LENGTH_LONG).show();
+//            }
+//        });
+//
+//        Button back1 = (Button) findViewById(R.id.back1);
+//
+//        back1.setOnClickListener(new View.OnClickListener() {
+//            @Override
+//            public void onClick(View v) {
+//                Intent back = new Intent(SdAvailable.this, ListActivity.class);
+//                startActivity(back);
 //
 //            }
 //        });
@@ -260,61 +278,10 @@ public class RCAvailable extends AppCompatActivity {
 //        back_button.setOnClickListener((new OnClickListener() {
 //            @Override
 //            public void onClick(View view) {
-//                Intent backIntent = new Intent(RCAvailable.this, ProfileActivity.class);
+//                Intent backIntent = new Intent(SdAvailable.this, ProfileActivity.class);
 //
 //                startActivity(backIntent);
 //            }
 //        }));
-//
-//        Button back1 = (Button) findViewById(R.id.back1);
-//
-//        back1.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                Intent back = new Intent(RCAvailable.this, ListActivity.class);
-//                startActivity(back);
-//
-//            }
-//        });
-//
 //    }
-
-
-
-
 }
-
-
-
-
-
-/*package com.example.toshiba.parkme;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.widget.Button;
-
-
-public class AvailableActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_available);
-
-        Button back_button = (Button) findViewById(R.id.back);
-
-        back_button.setOnClickListener((new OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent backIntent = new Intent(AvailableActivity.this, ProfileActivity.class);
-
-                startActivity(backIntent);
-            }
-        }));
-    }
-}
-*/
