+++ app/src/main/java/com/example/toshiba/parkme/activity/statusActivity.java
@@ -1,14 +1,16 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.activity;
 
 import android.app.Activity;
-import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
+
+import com.example.toshiba.parkme.R;
+import com.example.toshiba.parkme.model.User;
 
 import java.util.List;
 
@@ -18,7 +20,7 @@ public class statusActivity extends ArrayAdapter {
     private List<User> userinfo;
 
    public statusActivity(Activity context, List<User> userinfo){
-       super(context,R.layout.status_layout, userinfo);
+       super(context, R.layout.status_layout, userinfo);
 
        this.context = context;
        this.userinfo =userinfo;
