+++ app/src/main/java/com/example/toshiba/parkme/model/User.java
@@ -1,4 +1,4 @@
-package com.example.toshiba.parkme;
+package com.example.toshiba.parkme.model;
 
 public class User {
