+++ app/src/main/res/drawable/login_page.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient android:angle="270"
+        android:startColor="@color/white"
+        android:centerColor="@color/white"
+        android:endColor="@color/colorAccent"/>
+</shape>
