+++ app/src/main/res/layout/status_layout.xml
@@ -3,8 +3,9 @@
 <ScrollView
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.toshiba.parkme.SignupActivity"
+    tools:context="com.example.toshiba.parkme.activity.SignupActivity"
     android:layout_width="match_parent"
+    android:background="@drawable/login_page"
     android:layout_height="match_parent">
 <LinearLayout
     android:orientation="vertical"
